refactor: decompose cos.js dequeueNextTask and split cosAgents modules (#2530)
