+++ server/services/cosAgents.js
@@ -1,1093 +1,18 @@
 /**
- * CoS Agents Module
+ * CoS Agents Module (barrel)
  *
- * Agent lifecycle management extracted from cos.js.
- * Handles agent registration, state tracking, output capture,
- * date-bucketed archival, zombie cleanup, feedback, and index management.
+ * Agent lifecycle management, previously a single ~1090-line file. Decomposed
+ * (issue #2530) into focused sibling modules; this file stays as a thin barrel
+ * so every existing `from './cosAgents.js'` import path keeps working:
+ *
+ * - cosAgentIndex.js     — date-bucket index, migration, prune, archive layout
+ * - cosAgentLifecycle.js — register / update / complete / output / terminate /
+ *                          pause / kill / BTW / zombie cleanup / single reads
+ * - cosAgentFeedback.js  — per-agent feedback capture + aggregation + classifier
+ * - cosAgentArchive.js   — state-eviction sweeps (archive stale, clear completed)
  */
 
-import { readFile, writeFile, rename, readdir, rm, stat } from 'fs/promises';
-import { existsSync } from 'fs';
-import { join } from 'path';
-import { cosEvents, emitLog } from './cosEvents.js';
-import { loadState, saveState, withStateLock, AGENTS_DIR } from './cosState.js';
-import { atomicWrite, ensureDir, safeJSONParse, tryReadFile } from '../lib/fileUtils.js';
-import { recordDomainUsage } from './domainUsage.js';
-import { repairCodexTaskSummary } from './codexSummaryRepair.js';
-
-const INDEX_FILE = join(AGENTS_DIR, 'index.json');
-
-// Lightweight index mapping agentId → YYYY-MM-DD date bucket (~50KB vs 16MB full cache)
-// Lazy-loaded from data/cos/agents/index.json on first access
-let agentIndex = null;
-let agentIndexPromise = null;
-
-// Load agent index from disk (lazy init, singleton promise prevents concurrent migrations)
-export async function loadAgentIndex() {
-  if (agentIndex) return agentIndex;
-  if (agentIndexPromise) return agentIndexPromise;
-
-  agentIndexPromise = (async () => {
-    if (!existsSync(AGENTS_DIR)) {
-      await ensureDir(AGENTS_DIR);
-    }
-
-    if (existsSync(INDEX_FILE)) {
-      const content = await tryReadFile(INDEX_FILE);
-      const parsed = safeJSONParse(content ?? '{}', {});
-      agentIndex = new Map(Object.entries(parsed));
-      console.log(`📂 Loaded agent index: ${agentIndex.size} entries`);
-    } else {
-      // No index yet — run migration from flat dirs to date buckets
-      agentIndex = await migrateAgentsToDateBuckets();
-    }
-
-    return agentIndex;
-  })().catch(err => {
-    agentIndexPromise = null;
-    throw err;
-  });
-
-  return agentIndexPromise;
-}
-
-// Persist agent index to disk via the shared atomicWrite helper (temp file + rename,
-// with Windows backup-swap fallback). Without atomic semantics a mid-write crash
-// truncates index.json and on next boot the date-bucket migration would silently
-// re-run (or worse, drop already-archived agents from the lookup).
-async function saveAgentIndex() {
-  if (!agentIndex) return;
-  const obj = Object.fromEntries(agentIndex);
-  await atomicWrite(INDEX_FILE, obj).catch(err => {
-    console.error(`❌ Failed to save agent index: ${err.message}`);
-  });
-}
-
-/**
- * Merge a batch of {agentId, date} pairs into the agent index. Used by the
- * peer-sync CoS-history receiver (#1650): after pulling a peer's completed-agent
- * archives onto local disk, the date-bucket index must learn the new agentIds so
- * the history UI lists them. Idempotent union — only valid YYYY-MM-DD-bucketed
- * pairs are accepted. Returns the number of NEW entries added.
- *
- * A pre-existing agentId is treated as already-owned and is NEVER overwritten,
- * even when the incoming `date` differs: agent ids are generated independently
- * per instance, so an id collision (or a restored local archive) must not be able
- * to repoint a local agent's bucket at a peer's date — that would make
- * `getAgent()` read the wrong (or non-existent) directory and hide the local
- * agent's own history. First write wins; the local entry is authoritative.
- */
-export async function addAgentArchivesToIndex(pairs) {
-  if (!Array.isArray(pairs) || pairs.length === 0) return 0;
-  const idx = await loadAgentIndex();
-  let added = 0;
-  for (const pair of pairs) {
-    const agentId = pair?.agentId;
-    const date = pair?.date;
-    if (typeof agentId !== 'string' || !agentId) continue;
-    if (typeof date !== 'string' || !/^\d{4}-\d{2}-\d{2}$/.test(date)) continue;
-    if (idx.has(agentId)) continue; // already owned — never overwrite (id is authoritative)
-    idx.set(agentId, date);
-    added += 1;
-  }
-  if (added > 0) await saveAgentIndex();
-  return added;
-}
-
-// Resolve the correct directory for an agent (running = flat, completed = date bucket)
-function getAgentDir(agentId, dateString) {
-  if (dateString) return join(AGENTS_DIR, dateString, agentId);
-  // Check index for date bucket
-  const date = agentIndex?.get(agentId);
-  if (date) return join(AGENTS_DIR, date, agentId);
-  // Fallback to flat dir (running agents or pre-migration)
-  return join(AGENTS_DIR, agentId);
-}
-
-// Migrate flat agent-* directories into YYYY-MM-DD date buckets
-// Runs once when index.json doesn't exist. Idempotent — no-op if already migrated.
-async function migrateAgentsToDateBuckets() {
-  const index = new Map();
-
-  if (!existsSync(AGENTS_DIR)) {
-    await ensureDir(AGENTS_DIR);
-    await atomicWrite(INDEX_FILE, {});
-    console.log('📂 Created empty agent index (no agents to migrate)');
-    return index;
-  }
-
-  const entries = await readdir(AGENTS_DIR, { withFileTypes: true });
-
-  // Also scan existing date-bucket dirs to include them in the index
-  const dateDirPattern = /^\d{4}-\d{2}-\d{2}$/;
-  for (const entry of entries) {
-    if (!entry.isDirectory() || !dateDirPattern.test(entry.name)) continue;
-    const dateStr = entry.name;
-    const dateDir = join(AGENTS_DIR, dateStr);
-    const agentDirs = await readdir(dateDir, { withFileTypes: true }).catch(() => []);
-    for (const agentEntry of agentDirs) {
-      if (agentEntry.isDirectory() && agentEntry.name.startsWith('agent-')) {
-        index.set(agentEntry.name, dateStr);
-      }
-    }
-  }
-
-  // Find flat agent-* dirs that need migration
-  const flatAgentDirs = entries.filter(e => e.isDirectory() && e.name.startsWith('agent-'));
-
-  if (flatAgentDirs.length === 0) {
-    await atomicWrite(INDEX_FILE, Object.fromEntries(index));
-    console.log(`📂 Agent index built: ${index.size} entries (no flat dirs to migrate)`);
-    return index;
-  }
-
-  console.log(`📦 Migrating ${flatAgentDirs.length} agents into date buckets...`);
-  let migrated = 0;
-  let skipped = 0;
-
-  for (const entry of flatAgentDirs) {
-    const agentId = entry.name;
-    const agentDir = join(AGENTS_DIR, agentId);
-    const metaPath = join(agentDir, 'metadata.json');
-
-    let dateStr = null;
-
-    // Try to get date from metadata
-    if (existsSync(metaPath)) {
-      const content = await tryReadFile(metaPath);
-      if (content) {
-        const raw = safeJSONParse(content, null);
-        if (raw?.completedAt) {
-          dateStr = raw.completedAt.slice(0, 10); // YYYY-MM-DD
-        }
-      }
-    }
-
-    // Fallback: directory mtime
-    if (!dateStr) {
-      const dirStat = await stat(agentDir).catch(() => null);
-      if (dirStat?.mtime) {
-        dateStr = dirStat.mtime.toISOString().slice(0, 10);
-      }
-    }
-
-    if (!dateStr) {
-      console.log(`⚠️ Cannot determine date for ${agentId}, skipping`);
-      skipped++;
-      continue;
-    }
-
-    // Move into date bucket
-    const bucketDir = join(AGENTS_DIR, dateStr);
-    await ensureDir(bucketDir);
-    const targetDir = join(bucketDir, agentId);
-
-    // If target already exists (partial previous migration), skip
-    if (existsSync(targetDir)) {
-      index.set(agentId, dateStr);
-      migrated++;
-      continue;
-    }
-
-    await rename(agentDir, targetDir).catch(async (renameErr) => {
-      // rename can fail across filesystems — fall back to copy+delete
-      console.log(`⚠️ Rename failed for ${agentId}, using copy: ${renameErr.message}`);
-      try {
-        await ensureDir(targetDir);
-        const files = await readdir(agentDir);
-        for (const file of files) {
-          const content = await readFile(join(agentDir, file));
-          await writeFile(join(targetDir, file), content);
-        }
-        await rm(agentDir, { recursive: true });
-      } catch (copyErr) {
-        console.error(`❌ Copy fallback failed for ${agentId}: ${copyErr.message}`);
-        // Clean up partially-created target to avoid skipping on next startup
-        await rm(targetDir, { recursive: true, force: true }).catch(() => {});
-        throw copyErr;
-      }
-    });
-
-    index.set(agentId, dateStr);
-    migrated++;
-  }
-
-  // Persist index
-  await atomicWrite(INDEX_FILE, Object.fromEntries(index));
-  const uniqueDates = new Set(index.values()).size;
-  const parts = [`📦 Migrated ${migrated} agents into date buckets (${uniqueDates} unique dates)`];
-  if (skipped > 0) parts.push(`skipped ${skipped} undatable`);
-  console.log(parts.join(', '));
-
-  return index;
-}
-
-// Prune agent archive date buckets older than retentionDays (default 90).
-// Removes directories + their index entries. Runs after migration on startup.
-export async function pruneOldAgentArchives(retentionDays = 90) {
-  const idx = await loadAgentIndex();
-  if (!idx || idx.size === 0) return;
-
-  const cutoff = new Date();
-  cutoff.setDate(cutoff.getDate() - retentionDays);
-  const cutoffStr = cutoff.toISOString().slice(0, 10);
-
-  const dateDirPattern = /^\d{4}-\d{2}-\d{2}$/;
-  const entries = await readdir(AGENTS_DIR, { withFileTypes: true }).catch(() => []);
-  const oldDates = entries
-    .filter(e => e.isDirectory() && dateDirPattern.test(e.name) && e.name < cutoffStr)
-    .map(e => e.name);
-
-  if (oldDates.length === 0) return;
-
-  for (const dateStr of oldDates) {
-    await rm(join(AGENTS_DIR, dateStr), { recursive: true }).catch(() => {});
-  }
-
-  // Remove index entries for all old dates in a single pass
-  const oldDateSet = new Set(oldDates);
-  let pruned = 0;
-  for (const [agentId, date] of idx.entries()) {
-    if (oldDateSet.has(date)) { idx.delete(agentId); pruned++; }
-  }
-
-  await saveAgentIndex();
-  console.log(`🗑️ Pruned ${pruned} archived agents older than ${retentionDays} days (${oldDates.length} date buckets)`);
-}
-
-export async function registerAgent(agentId, taskId, metadata = {}) {
-  return withStateLock(async () => {
-    const state = await loadState();
-
-    state.agents[agentId] = {
-      id: agentId,
-      taskId,
-      status: 'running',
-      startedAt: new Date().toISOString(),
-      metadata,
-      output: []
-    };
-
-    state.stats.agentsSpawned++;
-    await saveState(state);
-
-    cosEvents.emit('agent:spawned', state.agents[agentId]);
-    return state.agents[agentId];
-  });
-}
-
-export async function updateAgent(agentId, updates) {
-  return withStateLock(async () => {
-    const state = await loadState();
-
-    if (!state.agents[agentId]) {
-      return null;
-    }
-
-    // Merge metadata if present in updates
-    if (updates.metadata) {
-      state.agents[agentId] = {
-        ...state.agents[agentId],
-        ...updates,
-        metadata: { ...state.agents[agentId].metadata, ...updates.metadata }
-      };
-    } else {
-      state.agents[agentId] = { ...state.agents[agentId], ...updates };
-    }
-    await saveState(state);
-
-    cosEvents.emit('agent:updated', state.agents[agentId]);
-    return state.agents[agentId];
-  });
-}
-
-export async function completeAgent(agentId, result = {}) {
-  const completed = await withStateLock(async () => {
-    const state = await loadState();
-
-    if (!state.agents[agentId]) {
-      return null;
-    }
-
-    state.agents[agentId] = {
-      ...state.agents[agentId],
-      status: 'completed',
-      completedAt: new Date().toISOString(),
-      // Success-criteria validation verdict (issue #2344): normalize to an
-      // explicit null sentinel when a completion path didn't declare/evaluate a
-      // machine-checkable criterion, so persisted telemetry never conflates
-      // "not declared" with "declared and failed" (false). Distinct from
-      // result.success (the runner's exit-code verdict).
-      result: { validationPassed: null, ...result },
-    };
-
-    if (result.success) {
-      state.stats.tasksCompleted++;
-    } else {
-      state.stats.errors = (state.stats.errors || 0) + 1;
-    }
-
-    await saveState(state);
-    // `agent:completed` is intentionally emitted later, after the domain-usage
-    // ledger is updated (see the recordDomainUsage block below, #1683). Do NOT
-    // move it back here. `agent:updated` carries no scheduling side effect, so it
-    // stays inside the lock.
-    cosEvents.emit('agent:updated', state.agents[agentId]);
-
-    // Determine date bucket from completedAt
-    const dateStr = state.agents[agentId].completedAt.slice(0, 10);
-    const bucketDir = join(AGENTS_DIR, dateStr);
-    await ensureDir(bucketDir);
-
-    // Write metadata to flat dir first (may already have output.txt/prompt.txt there)
-    const flatDir = join(AGENTS_DIR, agentId);
-    if (!existsSync(flatDir)) {
-      await ensureDir(flatDir);
-    }
-    const { output: _output, ...agentWithoutOutput } = state.agents[agentId];
-    await atomicWrite(join(flatDir, 'metadata.json'), agentWithoutOutput);
-
-    // Move entire agent dir into date bucket (atomic on same filesystem)
-    const targetDir = join(bucketDir, agentId);
-    if (!existsSync(targetDir)) {
-      await rename(flatDir, targetDir).catch(async () => {
-        // Fallback for cross-filesystem: copy files then remove
-        await ensureDir(targetDir);
-        const files = await readdir(flatDir);
-        for (const file of files) {
-          const content = await readFile(join(flatDir, file));
-          await writeFile(join(targetDir, file), content);
-        }
-        await rm(flatDir, { recursive: true });
-      });
-    }
-
-    // Update index
-    const idx = await loadAgentIndex();
-    idx.set(agentId, dateStr);
-    await saveAgentIndex();
-
-    return state.agents[agentId];
-  });
-
-  // Daily CoS budget accounting (#711): count only AUTONOMOUS runs (non-user
-  // tasks) — the same set the CoS auto-run gate withholds when over budget —
-  // toward the domain's actions/minutes ledger. Recorded outside the state lock
-  // (separate ledger file + write tail) so it never serializes against state.json.
-  //
-  // This MUST land before the `agent:completed` emit below: that event's handler
-  // schedules `dequeueNextTask()`, whose daily action-budget gate reads this
-  // ledger. Recording first ensures the gate counts the just-finished action, so
-  // a perpetual drain can't admit one spawn past `maxActionsPerDay` at the
-  // boundary (#1683).
-  if (completed?.metadata?.taskType && completed.metadata.taskType !== 'user') {
-    await recordDomainUsage('cos', { actions: 1, ms: Number(result.duration) || 0 })
-      .catch(err => console.error(`❌ Failed to record CoS budget usage for ${agentId}: ${err.message}`));
-  }
-
-  // Emit now that the ledger reflects this action (#1683). Fires for every
-  // completed agent, matching prior behavior — only the timing moved.
-  if (completed) {
-    cosEvents.emit('agent:completed', completed);
-  }
-
-  return completed;
-}
-
-export async function appendAgentOutput(agentId, line) {
-  const result = await withStateLock(async () => {
-    const state = await loadState();
-
-    if (!state.agents[agentId]) {
-      return null;
-    }
-
-    state.agents[agentId].output.push({
-      timestamp: new Date().toISOString(),
-      line
-    });
-
-    // Trim to last 1000 lines in state
-    if (state.agents[agentId].output.length > 1000) {
-      state.agents[agentId].output = state.agents[agentId].output.slice(-1000);
-    }
-
-    await saveState(state);
-    return state.agents[agentId];
-  });
-
-  if (result) {
-    cosEvents.emit('agent:output', { agentId, line });
-  }
-
-  return result;
-}
-
-// Batched variant — single state load+save for many lines. Used by the TUI
-// spawner to avoid write-amplification on chatty TUIs that emit hundreds of
-// lines per second; per-line appendAgentOutput would re-load and re-save the
-// entire state JSON for every line.
-export async function appendAgentOutputLines(agentId, lines) {
-  if (!Array.isArray(lines) || lines.length === 0) return null;
-  const result = await withStateLock(async () => {
-    const state = await loadState();
-    if (!state.agents[agentId]) return null;
-    const timestamp = new Date().toISOString();
-    for (const line of lines) {
-      state.agents[agentId].output.push({ timestamp, line });
-    }
-    if (state.agents[agentId].output.length > 1000) {
-      state.agents[agentId].output = state.agents[agentId].output.slice(-1000);
-    }
-    await saveState(state);
-    return state.agents[agentId];
-  });
-
-  if (result) {
-    for (const line of lines) {
-      cosEvents.emit('agent:output', { agentId, line });
-    }
-  }
-
-  return result;
-}
-
-// Debounce window for batching streamed agent output to state. A chatty
-// producer (CoS Runner stream events, non-stream CLI stdout) can emit dozens
-// of lines/sec; without batching, each line round-trips a full state
-// load+save (see appendAgentOutput). 250ms is invisible to the live tail but
-// cuts state I/O by 1-2 orders of magnitude. Matches OUTPUT_FLUSH_INTERVAL_MS
-// in agentTuiSpawning.js.
-const OUTPUT_FLUSH_INTERVAL_MS = 250;
-
-// Debounced per-agent output batcher. Wraps appendAgentOutputLines with a
-// ~250ms flush window so a hot streaming producer triggers one state load+save
-// per window instead of per line — the write-amplification guard documented in
-// CLAUDE.md ("High-frequency state writes must batch"). The TUI spawner rolls
-// its own equivalent inline because it co-flushes an output.txt appendFile in
-// the same batch; producers that only need the state write should use this.
-//
-// Callers MUST `await flush()` in their finish/cleanup path before the
-// completion event so the final lines land before the agent is marked done.
-export function createAgentOutputBatcher(agentId, { intervalMs = OUTPUT_FLUSH_INTERVAL_MS } = {}) {
-  let pending = [];
-  let timer = null;
-  let flushing = null;
-
-  const drain = async () => {
-    if (timer) { clearTimeout(timer); timer = null; }
-    if (pending.length === 0) return;
-    const batch = pending;
-    pending = [];
-    // Swallow+log state-write failures so neither the debounced timer nor a
-    // caller's `await flush()` ever rejects — this runs in child-process /
-    // timer callbacks where an uncaught throw would crash Node (CLAUDE.md "No
-    // try/catch" exception). The authoritative transcript lives in output.txt;
-    // a dropped live-tail batch is non-fatal. Mirrors the TUI spawner's
-    // `.catch(() => {})` on its own batched append.
-    await appendAgentOutputLines(agentId, batch).catch((err) => {
-      console.error(`❌ agent ${agentId} output batch flush failed: ${err.message}`);
-    });
-  };
-
-  const schedule = () => {
-    if (timer || flushing) return;
-    timer = setTimeout(() => {
-      timer = null;
-      flushing = drain().finally(() => {
-        flushing = null;
-        // Catch lines that arrived during the in-flight drain — without this a
-        // producer that goes quiet right after the timer fires strands its last
-        // batch in `pending` until flush().
-        if (pending.length > 0) schedule();
-      });
-    }, intervalMs);
-  };
-
-  return {
-    push(lineOrLines) {
-      if (Array.isArray(lineOrLines)) {
-        if (lineOrLines.length === 0) return;
-        pending.push(...lineOrLines);
-      } else {
-        pending.push(lineOrLines);
-      }
-      schedule();
-    },
-    // Wait for any in-flight drain, then fully empty `pending`. A push can race
-    // in during an awaited drain, so loop until nothing is left rather than
-    // draining a fixed number of times (which could strand a late line to the
-    // debounce timer). flush() is only called once the producer has stopped, so
-    // the loop terminates promptly.
-    async flush() {
-      if (flushing) await flushing;
-      while (pending.length > 0) await drain();
-    },
-  };
-}
-
-// Get all agents from in-memory state (includes running and recently completed; archived agents loaded via getAgentsByDate)
-export async function getAgents() {
-  const state = await loadState();
-  return Object.values(state.agents);
-}
-
-// Get available agent date buckets with counts, sorted descending
-export async function getAgentDates() {
-  const idx = await loadAgentIndex();
-  const dateCounts = {};
-  for (const date of idx.values()) {
-    dateCounts[date] = (dateCounts[date] || 0) + 1;
-  }
-  return Object.entries(dateCounts)
-    .map(([date, count]) => ({ date, count }))
-    .sort((a, b) => b.date.localeCompare(a.date));
-}
-
-// Get completed agents for a specific date bucket
-export async function getAgentsByDate(date) {
-  const dateDir = join(AGENTS_DIR, date);
-  if (!existsSync(dateDir)) return [];
-
-  const entries = await readdir(dateDir, { withFileTypes: true });
-  const agentDirs = entries.filter(e => e.isDirectory() && e.name.startsWith('agent-'));
-  const agents = [];
-
-  // Batch reads in chunks of 50 to avoid fd exhaustion on large date buckets
-  const BATCH_SIZE = 50;
-  for (let i = 0; i < agentDirs.length; i += BATCH_SIZE) {
-    const batch = agentDirs.slice(i, i + BATCH_SIZE);
-    const reads = batch.map(async (entry) => {
-      const metaPath = join(dateDir, entry.name, 'metadata.json');
-      const content = await tryReadFile(metaPath);
-      if (!content) return;
-      const raw = safeJSONParse(content, null);
-      if (!raw) return;
-      const id = raw.id || raw.agentId || entry.name;
-      const { output, ...rest } = raw;
-      const agent = { ...rest, id, status: raw.status || 'completed' };
-      const repaired = await repairCodexTaskSummary(join(dateDir, entry.name), agent);
-      if (repaired) agent.metadata = { ...agent.metadata, taskSummary: repaired };
-      agents.push(agent);
-    });
-    await Promise.allSettled(reads);
-  }
-
-  return agents.sort((a, b) => new Date(b.completedAt || 0) - new Date(a.completedAt || 0));
-}
-
-// Get agent by ID with full output from file
-export async function getAgent(agentId) {
-  const state = await loadState();
-  let agent = state.agents[agentId];
-
-  // Fall back to disk metadata via index if not in state
-  if (!agent) {
-    const idx = await loadAgentIndex();
-    const dateStr = idx.get(agentId);
-    if (dateStr) {
-      const metaPath = join(AGENTS_DIR, dateStr, agentId, 'metadata.json');
-      const content = await tryReadFile(metaPath);
-      if (content) {
-        const raw = safeJSONParse(content, null);
-        if (raw) {
-          const { output, ...rest } = raw;
-          agent = { ...rest, id: raw.id || raw.agentId || agentId, status: raw.status || 'completed' };
-        }
-      }
-    }
-  }
-  if (!agent) return null;
-
-  // Completed agents live in date buckets; paused agents remain in the flat
-  // agent dir but should still expose their preserved full transcript.
-  if (agent.status === 'completed' || agent.status === 'paused') {
-    const dateStr = agent.status === 'completed' ? agent.completedAt?.slice(0, 10) : null;
-    const agentDir = dateStr ? getAgentDir(agentId, dateStr) : getAgentDir(agentId);
-    const repaired = await repairCodexTaskSummary(agentDir, agent);
-    if (repaired) agent = { ...agent, metadata: { ...agent.metadata, taskSummary: repaired } };
-    const outputFile = join(agentDir, 'output.txt');
-    if (existsSync(outputFile)) {
-      const fullOutput = await readFile(outputFile, 'utf-8');
-      const lines = fullOutput.split('\n').filter(line => line.trim());
-      const timestamp = agent.completedAt || agent.pausedAt;
-      return {
-        ...agent,
-        output: lines.map(line => ({ line, timestamp }))
-      };
-    }
-  }
-
-  return agent;
-}
-
-// Read the prompt that was sent to an agent at spawn time.
-// Used by the AgentCard UI to let the user inspect what was pasted into the
-// TUI / sent to the CLI so the prompt can be iterated on.
-export async function getAgentPrompt(agentId) {
-  const state = await loadState();
-  const agent = state.agents[agentId];
-  if (!agent) return { error: 'Agent not found' };
-  const agentDir = getAgentDir(agentId, agent.archiveDate);
-  const promptPath = join(agentDir, 'prompt.txt');
-  if (!existsSync(promptPath)) return { error: 'Prompt file not found' };
-  const prompt = await readFile(promptPath, 'utf8');
-  return { prompt, bytes: prompt.length };
-}
-
-// Terminate an agent (will be handled by spawner)
-export async function terminateAgent(agentId) {
-  // Emit event to kill the process FIRST
-  cosEvents.emit('agent:terminate', agentId);
-  // The spawner will handle marking the agent as completed after termination
-  return { success: true, agentId };
-}
-
-// Pause an agent without completing its task or cleaning up its worktree.
-export async function pauseAgent(agentId, reason = null) {
-  const { pauseAgent: pauseAgentFromSpawner } = await import('./subAgentSpawner.js');
-  return pauseAgentFromSpawner(agentId, reason);
-}
-
-// Force kill an agent with SIGKILL (immediate, no graceful shutdown)
-export async function killAgent(agentId) {
-  const { killAgent: killAgentFromSpawner } = await import('./subAgentSpawner.js');
-  return killAgentFromSpawner(agentId);
-}
-
-// Send a BTW (additional context) message to a running agent.
-//
-// BTW is only supported for Claude Code TUI agents — the message gets
-// bracket-pasted directly into the live PTY session as if the user typed it
-// themselves. The legacy BTW.md path is gone: it required headless agents to
-// poll a file mid-run, which most CLIs (codex / antigravity / LM Studio) don't do
-// reliably anyway, and the indirection had to be reflected in the prompt with
-// a brittle "check this file" instruction. Other TUI kinds (codex, antigravity)
-// don't honor bracketed-paste in the same way, so they're not eligible
-// either.
-export async function sendBtwToAgent(agentId, message) {
-  const agentInfo = await withStateLock(async () => {
-    const state = await loadState();
-    const agent = state.agents[agentId];
-    if (!agent) return { error: 'Agent not found' };
-    if (agent.status !== 'running') return { error: 'Agent is not running' };
-    if (agent.metadata?.executionMode !== 'tui') {
-      return { error: 'BTW is only supported for Claude Code TUI agents.' };
-    }
-    if (agent.metadata?.tuiKind !== 'claude') {
-      return { error: `BTW is only supported for Claude Code TUI agents (this agent runs ${agent.metadata.tuiKind || 'an unknown TUI'}).` };
-    }
-    if (!agent.metadata?.tuiSessionId) {
-      return { error: 'Agent has no attached TUI session' };
-    }
-    return { tuiSessionId: agent.metadata.tuiSessionId };
-  });
-
-  if (agentInfo.error) return agentInfo;
-
-  const shellService = await import('./shell.js');
-  if (!shellService.getSession(agentInfo.tuiSessionId)) {
-    return { error: 'TUI session is no longer alive' };
-  }
-  // Bracketed-paste + delayed Enter, mirroring the initial prompt paste in
-  // agentTuiSpawning.js: Claude Code commits the paste buffer before the
-  // submit arrives, so multi-line messages land as a single paste event.
-  shellService.writeToSession(agentInfo.tuiSessionId, `\x1b[200~${message}\x1b[201~`);
-  setTimeout(() => {
-    try {
-      // Re-check session liveness: the TUI session may have died in the 400ms window.
-      if (shellService.getSession(agentInfo.tuiSessionId)) {
-        shellService.writeToSession(agentInfo.tuiSessionId, '\r');
-      }
-    } catch (err) {
-      console.error(`❌ [cosAgents] BTW delayed Enter failed for session ${agentInfo.tuiSessionId}: ${err.message}`);
-    }
-  }, 400);
-
-  // Track in agent state (cap at 50 messages)
-  const timestamp = new Date().toISOString();
-  await withStateLock(async () => {
-    const state = await loadState();
-    if (!state.agents[agentId]) return;
-    if (!state.agents[agentId].btwMessages) {
-      state.agents[agentId].btwMessages = [];
-    }
-    state.agents[agentId].btwMessages.push({ message, timestamp });
-    if (state.agents[agentId].btwMessages.length > 50) {
-      state.agents[agentId].btwMessages = state.agents[agentId].btwMessages.slice(-50);
-    }
-    await saveState(state);
-  });
-
-  cosEvents.emit('agent:btw', { agentId, message, timestamp });
-  return { success: true, delivered: 'tui-paste', tuiSessionId: agentInfo.tuiSessionId };
-}
-
-// Get process stats for an agent (CPU, memory)
-export async function getAgentProcessStats(agentId) {
-  const { getAgentProcessStats: getStatsFromSpawner } = await import('./subAgentSpawner.js');
-  return getStatsFromSpawner(agentId);
-}
-
-// Check if a PID is still running
-async function isPidAlive(pid) {
-  if (!pid) return false;
-  try {
-    process.kill(pid, 0);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-// Cleanup zombie agents - agents marked as running but whose process is dead
-export async function cleanupZombieAgents() {
-  // Check local tracking maps (read from the side-effect-free state module, not
-  // subAgentSpawner — avoids pulling in the heavier orchestrator just to read the maps).
-  const { getActiveAgentIds } = await import('./agentState.js');
-  const activeIds = getActiveAgentIds();
-
-  // Also check with the CoS runner for agents it's actively tracking
-  const { getActiveAgentsFromRunner } = await import('./cosRunnerClient.js');
-  const runnerAgents = await getActiveAgentsFromRunner().catch(() => []);
-  const runnerAgentIds = new Set(runnerAgents.map(a => a.id));
-
-  return withStateLock(async () => {
-    const state = await loadState();
-    const runningAgents = Object.values(state.agents).filter(a => a.status === 'running');
-    const cleaned = [];
-
-    for (const agent of runningAgents) {
-      // Skip if tracked in local maps or runner
-      if (activeIds.includes(agent.id) || runnerAgentIds.has(agent.id)) {
-        continue;
-      }
-
-      // If agent has a PID, verify the process is actually dead
-      if (agent.pid) {
-        const alive = await isPidAlive(agent.pid);
-        if (alive) continue;
-      } else {
-        // No PID yet - agent might still be initializing
-        // Give it a 30 second grace period before marking as zombie
-        const startedAt = agent.startedAt ? new Date(agent.startedAt).getTime() : 0;
-        const ageMs = Date.now() - startedAt;
-        if (ageMs < 30000) continue;
-      }
-
-      // Agent is not tracked anywhere and process is dead — it's a zombie
-      console.log(`🧟 Zombie agent detected: ${agent.id} (PID ${agent.pid || 'unknown'} not running)`);
-      state.agents[agent.id] = {
-        ...agent,
-        status: 'completed',
-        completedAt: new Date().toISOString(),
-        result: { success: false, error: 'Agent process terminated unexpectedly' }
-      };
-      cleaned.push(agent.id);
-    }
-
-    if (cleaned.length > 0) {
-      await saveState(state);
-
-      // Persist zombie-cleaned agents to date-bucketed dirs and update index
-      const idx = await loadAgentIndex();
-      for (const agentId of cleaned) {
-        const agent = state.agents[agentId];
-        const dateStr = agent.completedAt?.slice(0, 10);
-        if (!dateStr) continue;
-        const bucketDir = join(AGENTS_DIR, dateStr);
-        await ensureDir(bucketDir);
-
-        const flatDir = join(AGENTS_DIR, agentId);
-        const { output, ...agentWithoutOutput } = agent;
-
-        // Ensure metadata is written before move
-        if (!existsSync(flatDir)) await ensureDir(flatDir);
-        await atomicWrite(join(flatDir, 'metadata.json'), agentWithoutOutput).catch(() => {});
-
-        // Move to date bucket
-        const targetDir = join(bucketDir, agentId);
-        if (!existsSync(targetDir)) {
-          await rename(flatDir, targetDir).catch(async () => {
-            await ensureDir(targetDir);
-            const files = await readdir(flatDir);
-            for (const file of files) {
-              const content = await readFile(join(flatDir, file));
-              await writeFile(join(targetDir, file), content);
-            }
-            await rm(flatDir, { recursive: true });
-          });
-        }
-
-        idx.set(agentId, dateStr);
-      }
-      await saveAgentIndex();
-
-      console.log(`🧹 Cleaned up ${cleaned.length} zombie agents: ${cleaned.join(', ')}`);
-      cosEvents.emit('agents:changed', { action: 'zombie-cleanup', cleaned });
-    }
-
-    return { cleaned, count: cleaned.length };
-  });
-}
-
-// Delete a single agent from state and disk
-export async function deleteAgent(agentId) {
-  return withStateLock(async () => {
-    const state = await loadState();
-    const idx = await loadAgentIndex();
-
-    const inState = !!state.agents[agentId];
-    const inIndex = idx.has(agentId);
-    if (!inState && !inIndex) {
-      return { error: 'Agent not found' };
-    }
-
-    delete state.agents[agentId];
-    await saveState(state);
-
-    // Remove from disk (date-bucketed or flat)
-    const agentDir = getAgentDir(agentId);
-    if (existsSync(agentDir)) {
-      await rm(agentDir, { recursive: true }).catch(() => {});
-    }
-
-    // Remove from index
-    idx.delete(agentId);
-    await saveAgentIndex();
-
-    cosEvents.emit('agents:changed', { action: 'deleted', agentId });
-    return { success: true, agentId };
-  });
-}
-
-// Submit feedback for a completed agent
-export async function submitAgentFeedback(agentId, feedback) {
-  return withStateLock(async () => {
-    const state = await loadState();
-    const feedbackData = {
-      rating: feedback.rating,
-      comment: feedback.comment || null,
-      submittedAt: new Date().toISOString()
-    };
-
-    // Try state first (recently completed agents still in state)
-    if (state.agents[agentId]) {
-      const agent = state.agents[agentId];
-      if (agent.status !== 'completed') {
-        return { error: 'Can only submit feedback for completed agents' };
-      }
-      state.agents[agentId].feedback = feedbackData;
-      await saveState(state);
-
-      // Also update on-disk metadata (derive date bucket from completedAt if archived)
-      const dateBucket = agent.completedAt ? agent.completedAt.slice(0, 10) : null;
-      const agentDir = getAgentDir(agentId, dateBucket);
-      const metaPath = join(agentDir, 'metadata.json');
-      if (existsSync(metaPath)) {
-        const content = await tryReadFile(metaPath);
-        if (content) {
-          const raw = safeJSONParse(content, null);
-          if (raw) {
-            raw.feedback = feedbackData;
-            await atomicWrite(metaPath, raw).catch(() => {});
-          }
-        }
-      }
-
-      emitLog('info', `Feedback received for agent ${agentId}: ${feedback.rating}`, { agentId, rating: feedback.rating });
-      cosEvents.emit('agent:feedback', { agentId, feedback: feedbackData });
-      return { success: true, agent: state.agents[agentId] };
-    }
-
-    // Agent not in state — look up from disk via index
-    const idx = await loadAgentIndex();
-    const dateStr = idx.get(agentId);
-    if (!dateStr) return { error: 'Agent not found' };
-
-    const metaPath = join(AGENTS_DIR, dateStr, agentId, 'metadata.json');
-    const content = await tryReadFile(metaPath);
-    if (!content) return { error: 'Agent not found' };
-
-    const raw = safeJSONParse(content, null);
-    if (!raw) return { error: 'Agent not found' };
-
-    raw.feedback = feedbackData;
-    await atomicWrite(metaPath, raw);
-
-    emitLog('info', `Feedback received for agent ${agentId}: ${feedback.rating}`, { agentId, rating: feedback.rating });
-    cosEvents.emit('agent:feedback', { agentId, feedback: feedbackData });
-    return { success: true, agent: { ...raw, id: agentId } };
-  });
-}
-
-// Get aggregated feedback statistics
-export async function getFeedbackStats() {
-  const state = await loadState();
-  const agents = Object.values(state.agents);
-
-  const withFeedback = agents.filter(a => a.feedback);
-  const positive = withFeedback.filter(a => a.feedback.rating === 'positive').length;
-  const negative = withFeedback.filter(a => a.feedback.rating === 'negative').length;
-  const neutral = withFeedback.filter(a => a.feedback.rating === 'neutral').length;
-
-  // Group by task type
-  const byTaskType = {};
-  withFeedback.forEach(a => {
-    const taskType = extractTaskType(a.metadata?.taskDescription);
-    if (!byTaskType[taskType]) {
-      byTaskType[taskType] = { positive: 0, negative: 0, neutral: 0, total: 0 };
-    }
-    byTaskType[taskType][a.feedback.rating]++;
-    byTaskType[taskType].total++;
-  });
-
-  // Recent feedback (last 10 with comments)
-  const recentWithComments = withFeedback
-    .filter(a => a.feedback.comment)
-    .sort((a, b) => new Date(b.feedback.submittedAt) - new Date(a.feedback.submittedAt))
-    .slice(0, 10)
-    .map(a => ({
-      agentId: a.id,
-      taskDescription: a.metadata?.taskDescription,
-      rating: a.feedback.rating,
-      comment: a.feedback.comment,
-      submittedAt: a.feedback.submittedAt
-    }));
-
-  const satisfactionRate = withFeedback.length > 0
-    ? Math.round((positive / withFeedback.length) * 100)
-    : null;
-
-  return {
-    total: withFeedback.length,
-    positive,
-    negative,
-    neutral,
-    satisfactionRate,
-    byTaskType,
-    recentWithComments
-  };
-}
-
-// Helper to extract task type from description (mirrors client-side logic)
-export function extractTaskType(description) {
-  if (!description) return 'general';
-  const d = description.toLowerCase();
-  if (d.includes('fix') || d.includes('bug') || d.includes('error') || d.includes('issue')) return 'bug-fix';
-  if (d.includes('refactor') || d.includes('clean up') || d.includes('improve') || d.includes('optimize')) return 'refactor';
-  if (d.includes('test')) return 'testing';
-  if (d.includes('document') || d.includes('readme') || d.includes('docs')) return 'documentation';
-  if (d.includes('review') || d.includes('audit')) return 'code-review';
-  if (d.includes('mobile') || d.includes('responsive')) return 'mobile-responsive';
-  if (d.includes('security') || d.includes('vulnerability')) return 'security';
-  if (d.includes('performance') || d.includes('speed')) return 'performance';
-  if (d.includes('ui') || d.includes('ux') || d.includes('design') || d.includes('style')) return 'ui-ux';
-  if (d.includes('api') || d.includes('endpoint') || d.includes('route')) return 'api';
-  if (d.includes('database') || d.includes('migration')) return 'database';
-  if (d.includes('deploy') || d.includes('ci') || d.includes('cd')) return 'devops';
-  if (d.includes('investigate') || d.includes('debug')) return 'investigation';
-  if (d.includes('self-improvement') || d.includes('feature idea')) return 'self-improvement';
-  return 'feature';
-}
-
-// Archive stale completed agents from state.json.
-// Completed agents are already persisted to per-agent metadata files on disk
-// (metadata.json) by completeAgent(), so removing them from state.json only
-// reduces the size of the in-memory state and the state.json file.
-export async function archiveStaleAgents() {
-  return withStateLock(async () => {
-    const state = await loadState();
-    const retentionMs = state.config.completedAgentRetentionMs ?? 86400000;
-    const cutoff = Date.now() - retentionMs;
-
-    const staleIds = Object.keys(state.agents).filter(id => {
-      const agent = state.agents[id];
-      if (agent.status !== 'completed') return false;
-      const completedAt = agent.completedAt ? new Date(agent.completedAt).getTime() : 0;
-      return completedAt > 0 && completedAt < cutoff;
-    });
-
-    if (staleIds.length === 0) return { archived: 0 };
-
-    const idx = await loadAgentIndex();
-
-    for (const id of staleIds) {
-      // Ensure agent is persisted to date-bucketed disk before removing from state
-      if (!idx.has(id)) {
-        const agent = state.agents[id];
-        const dateStr = agent.completedAt?.slice(0, 10);
-        if (!dateStr) continue;
-        const bucketDir = join(AGENTS_DIR, dateStr);
-        await ensureDir(bucketDir);
-
-        const { output, ...agentWithoutOutput } = agent;
-        const flatDir = join(AGENTS_DIR, id);
-        const targetDir = join(bucketDir, id);
-
-        if (existsSync(flatDir) && !existsSync(targetDir)) {
-          // Write metadata then move (with cross-filesystem fallback)
-          await atomicWrite(join(flatDir, 'metadata.json'), agentWithoutOutput).catch(() => {});
-          await rename(flatDir, targetDir).catch(async () => {
-            await ensureDir(targetDir);
-            const files = await readdir(flatDir).catch(() => []);
-            for (const file of files) {
-              const content = await tryReadFile(join(flatDir, file), null);
-              if (content !== null) await writeFile(join(targetDir, file), content);
-            }
-            await rm(flatDir, { recursive: true }).catch(() => {});
-          });
-          if (!existsSync(targetDir)) continue; // Skip index update if move failed
-        } else if (!existsSync(targetDir)) {
-          await ensureDir(targetDir);
-          await atomicWrite(join(targetDir, 'metadata.json'), agentWithoutOutput).catch(() => {});
-        }
-
-        idx.set(id, dateStr);
-      }
-
-      delete state.agents[id];
-    }
-
-    await saveState(state);
-    await saveAgentIndex();
-    console.log(`📦 Archived ${staleIds.length} stale agents from state.json (retained on disk)`);
-    cosEvents.emit('agents:changed', { action: 'auto-archive', archived: staleIds.length });
-    return { archived: staleIds.length };
-  });
-}
-
-// Clear completed agents from state, cache, and disk
-export async function clearCompletedAgents() {
-  return withStateLock(async () => {
-    const state = await loadState();
-    const idx = await loadAgentIndex();
-
-    // Remove completed agents from state
-    const stateCompleted = Object.keys(state.agents).filter(
-      id => state.agents[id].status === 'completed'
-    );
-    for (const id of stateCompleted) {
-      delete state.agents[id];
-    }
-    await saveState(state);
-
-    // Collect all unique dates from index, then remove date bucket dirs
-    const dates = new Set(idx.values());
-    const totalCleared = idx.size + stateCompleted.filter(id => !idx.has(id)).length;
-
-    const removals = [...dates].map(date => {
-      const dateDir = join(AGENTS_DIR, date);
-      return existsSync(dateDir)
-        ? rm(dateDir, { recursive: true }).catch(() => {})
-        : Promise.resolve();
-    });
-    await Promise.all(removals);
-
-    // Clear index
-    idx.clear();
-    await saveAgentIndex();
-
-    return { cleared: totalCleared };
-  });
-}
+export * from './cosAgentIndex.js';
+export * from './cosAgentLifecycle.js';
+export * from './cosAgentFeedback.js';
+export * from './cosAgentArchive.js';
